+++ rpc/Cargo.toml
@@ -3,6 +3,7 @@ name = "ckb-rpc"
 version = "0.3.0-pre"
 license = "MIT"
 authors = ["Nervos Core Dev <dev@nervos.org>"]
+edition = "2018"
 
 [dependencies]
 numext-fixed-hash = { version = "0.1", features = ["support_rand", "support_heapsize", "support_serde"] }
@@ -15,9 +16,9 @@ ckb-pool = { path = "../pool" }
 ckb-chain = { path = "../chain" }
 ckb-miner = { path = "../miner" }
 ckb-protocol = { path = "../protocol" }
-ckb-pow = { path = "../pow", optional = true }
+ckb-pow = { path = "../pow"}
 jsonrpc-core = "9.0"
-jsonrpc-macros = "9.0"
+jsonrpc-macros = { git = "https://github.com/nervosnetwork/jsonrpc.git", branch = "2018-edition" }
 jsonrpc-http-server = "9.0"
 jsonrpc-server-utils = "9.0"
 serde = "1.0"
@@ -25,10 +26,8 @@ serde_derive = "1.0"
 serde_json = "1.0"
 log = "0.4"
 flatbuffers = "0.5.0"
+num_cpus = "1.0"
 
 [dev-dependencies]
 ckb-db = { path = "../db" }
 ckb-verification = { path = "../verification" }
-
-[features]
-integration_test = ["ckb-pow"]
