+++ rpc/src/module/mod.rs
@@ -1,0 +1,11 @@
+mod chain;
+mod miner;
+mod net;
+mod pool;
+mod test;
+
+pub(crate) use self::chain::{ChainRpc, ChainRpcImpl};
+pub(crate) use self::miner::{MinerRpc, MinerRpcImpl};
+pub(crate) use self::net::{NetworkRpc, NetworkRpcImpl};
+pub(crate) use self::pool::{PoolRpc, PoolRpcImpl};
+pub(crate) use self::test::{IntegrationTestRpc, IntegrationTestRpcImpl};
