+++ rpc/src/types.rs
@@ -3,6 +3,7 @@ use ckb_core::cell::CellStatus;
 use ckb_core::header::Header;
 use ckb_core::transaction::{Capacity, CellOutput, OutPoint, Transaction};
 use numext_fixed_hash::H256;
+use serde_derive::Serialize;
 
 #[derive(Serialize)]
 pub struct TransactionWithHash {
@@ -68,9 +69,4 @@ impl From<CellStatus> for CellWithStatus {
             status: status.to_string(),
         }
     }
-}
-
-#[derive(Clone, Debug, PartialEq, Deserialize)]
-pub struct Config {
-    pub listen_addr: String,
 }
